refactor : 게임과 관련된 클래스 생성 및 자동차 클래스명 변경
- goOrStop함수는 게임 유틸리티에 더 적합하여 해당 클래스로 이동
- 자동차 클래스 역시 유틸리티 클래스이므로 클래스 네이밍을 맞춘다.
